Moved NoteModel to main Models folder.
