+++ ecommerce/ecommerce/Models/NoteModel.swift
@@ -1,0 +1,32 @@
+//
+//  NoteModel.swift
+//  ecommerce
+//
+//  Created by Vincent Salinas on 7/11/22.
+//
+
+import Foundation
+
+class NoteModel {
+    var productID: Int
+    var username: String
+    var title: String
+    var date: Date
+    var body: String
+    
+    init(productID: Int, username: String, title: String, date: Date, body: String) {
+        self.productID = productID
+        self.username = username
+        self.title = title
+        self.date = date
+        self.body = body
+    }
+    
+    static func fetchNotes() -> [NoteModel] {
+        return [
+            NoteModel(productID: 0, username: "name", title: "title", date: Date(), body: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."),
+            NoteModel(productID: 0, username: "name", title: "title", date: Date(), body: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."),
+            NoteModel(productID: 0, username: "name", title: "title", date: Date(), body: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."),
+        ]
+    }
+}
